+++ src/main/java/com/dateplan/dateplan/domain/member/dto/login/LoginServiceRequest.java
@@ -1,4 +1,4 @@
-package com.dateplan.dateplan.domain.member.dto;
+package com.dateplan.dateplan.domain.member.dto.login;
 
 import lombok.Builder;
 import lombok.Getter;
